requirements.txt
#pip install -r requirements.txt

streamlit
openai
python-dotenv
cachetools
langchain
pydantic
beautifulsoup4
requests
fastapi
google-search-results
tiktoken